Add mixins and defineProperties
