+++ src/mapbox/layers/TrackerLayer.js
@@ -1,94 +1,40 @@
-import { unByKey } from 'ol/Observable';
 import { toLonLat, fromLonLat } from 'ol/proj';
-import { buffer, containsCoordinate } from 'ol/extent';
-import Layer from './Layer';
-import Tracker from '../../common/Tracker';
-import { timeSteps } from '../../common/trackerConfig';
+import CommonLayer from '../../common/layers/Layer';
+import TrackerMixin from '../../common/mixins/TrackerMixin';
 import { getResolution } from '../utils';
 
 /**
  * Responsible for loading tracker data.
- * Extended from Layer {@link https://react-spatial.geops.de/docjs.html react-spatial/layers/Layer}
- * @class
- * @inheritDoc
  * @param {Object} options
  * @param {boolean} options.useDelayStyle Set the delay style.
  * @param {string} options.onClick Callback function on feature click.
+ * @class
  * @private
  */
-class TrackerLayer extends Layer {
+class TrackerLayer extends TrackerMixin(CommonLayer) {
   constructor(options = {}) {
     super({
       ...options,
     });
 
-    // Array of ol events key. Be careful to not override this value in child classe.
-    this.olEventsKeys = [];
-
-    /**
-     * Cache object for trajectories drawn.
-     * @private
-     */
-    this.styleCache = {};
-
-    /**
-     * Time speed.
-     * @private
-     */
-    this.speed = 1;
-
-    /**
-     * Time used to display the trajectories.
-     * @private
-     */
-    this.currTime = new Date();
-
-    /**
-     * Keep track of the last time used to render trajectories.
-     * Useful when the speed increase.
-     * @private
-     */
-    this.lastUpdateTime = new Date();
-
-    /**
-     * Activate/deactivate pointer hover effect.
-     * @private
-     */
-    this.isHoverActive =
-      options.isHoverActive !== undefined ? options.isHoverActive : true;
-
-    /**
-     * Callback function when a user click on a vehicle.
-     * @private
-     */
-    this.clickCallbacks = [];
-
-    // Add click callback
-    if (options.onClick) {
-      this.onClick(options.onClick);
-    }
-
-    /**
-     * Custom property for duck typing since `instanceof` is not working
-     * when the instance was created on different bundles.
-     * @public
-     */
-    this.isTrackerLayer = true;
+    this.onMapMoveEnd = this.onMapMoveEnd.bind(this);
+    this.onMapMouseMove = this.onMapMouseMove.bind(this);
   }
 
   /**
-   * Initialize the layer and listen to feature clicks.
-   * @param {ol.map} map {@link https://openlayers.org/en/latest/apidoc/module-ol_Map-Map.html Map}
+   * Initialize the layer.
+   * @param {mapboxgl.map} map A Mapbox [Map](https://docs.mapbox.com/mapbox-gl-js/api/map/).
    * @private
    */
   init(map) {
     super.init(map);
+
     if (!this.map) {
       return;
     }
 
     const { width, height } = map.getCanvas();
-    this.tracker = new Tracker({
+    super.initTracker({
       width,
       height,
       getPixelFromCoordinate: (coord) => {
@@ -97,41 +43,6 @@ class TrackerLayer extends Layer {
         return [x, y];
       },
     });
-    this.tracker.setStyle((props, r) => this.style(props, r));
-
-    if (this.getVisible()) {
-      this.start();
-    }
-
-    this.visibilityRef = this.on('change:visible', (evt) => {
-      if (evt.target.getVisible()) {
-        this.start();
-      } else {
-        this.stop();
-      }
-    });
-  }
-
-  terminate() {
-    this.stop();
-    unByKey(this.visibilityRef);
-    if (this.tracker) {
-      this.tracker.destroy();
-      this.tracker = null;
-    }
-    super.terminate();
-  }
-
-  /**
-   * Get the duration before the next update.
-   * @private
-   */
-  getRefreshTimeInMs() {
-    const z = this.map.getZoom();
-    const roundedZoom = Math.round(z);
-    const timeStep = timeSteps[roundedZoom] || 25;
-    const nextTick = Math.max(25, timeStep / this.speed);
-    return nextTick;
   }
 
   /**
@@ -139,135 +50,44 @@ class TrackerLayer extends Layer {
    * @param {dateString | value} time
    */
   setCurrTime(time) {
-    const newTime = new Date(time);
-    this.currTime = newTime;
-    this.lastUpdateTime = new Date();
-    if (
-      !this.map.isMoving() &&
-      !this.map.isRotating() &&
-      !this.map.isZooming()
-    ) {
-      const canvas = this.map.getCanvas();
-      this.tracker.renderTrajectories(
-        this.currTime,
-        [canvas.width, canvas.height],
-        100,
-      );
-    }
+    const canvas = this.map.getCanvas();
+    super.setCurrTime(
+      time,
+      [canvas.width, canvas.height],
+      getResolution(this.map),
+      !this.map.isMoving() && !this.map.isRotating() && !this.map.isZooming(),
+    );
   }
 
   /**
-   * Set the speed.
-   * @param {number} speed
-   */
-  setSpeed(speed) {
-    this.speed = speed;
-    this.start();
-  }
-
-  /**
-   * Trackerlayer is started
+   * Start updating vehicles position.
    * @param {ol.map} map {@link https://openlayers.org/en/latest/apidoc/module-ol_Map-Map.html ol/Map}
    * @private
    */
   start() {
-    this.stop();
-    this.tracker.setVisible(true);
-    const canvas = this.map.getCanvas();
-    this.tracker.renderTrajectories(
-      this.currTime,
-      [canvas.width, canvas.height],
-      100,
-    );
-    this.startUpdateTime();
     this.currentZoom = this.map.getZoom();
+    const canvas = this.map.getCanvas();
+    super.start(
+      [canvas.width, canvas.height],
+      this.currentZoom,
+      getResolution(this.map),
+    );
 
-    this.olEventsKeys = [
-      this.map.on('moveend', () => {
-        const z = this.map.getZoom();
+    this.map.on('moveend', this.onMapMoveEnd);
 
-        if (z !== this.currentZoom) {
-          this.currentZoom = z;
-          this.startUpdateTime();
-        }
-      }),
-      this.map.on('mousemove', (evt) => {
-        if (
-          this.map.isMoving() ||
-          this.map.isRotating() ||
-          this.map.isZooming() ||
-          !this.isHoverActive
-        ) {
-          this.map.getContainer().style.cursor = 'auto';
-          return;
-        }
-        const [vehicle] = this.getVehiclesAtCoordinate(
-          fromLonLat([evt.lngLat.lng, evt.lngLat.lat]),
-        );
-        this.map.getContainer().style.cursor = vehicle ? 'pointer' : 'auto';
-        this.tracker.setHoverVehicleId(vehicle && vehicle.id);
-      }),
-    ];
+    if (this.isHoverActive) {
+      this.map.on('mousemove', this.onMapMouseMove);
+    }
   }
 
   /**
-   * Stop current layer,.
+   * Stop updating vehicles position.
    * @private
    */
   stop() {
-    this.stopUpdateTime();
-    if (this.tracker) {
-      this.tracker.setVisible(false);
-      this.tracker.clear();
-    }
-    unByKey(this.olEventsKeys);
-    this.olEventsKeys = [];
-  }
-
-  /**
-   * Set the filter for tracker features.
-   * @param {Function} filter Filter function.
-   */
-  setFilter(filter) {
-    if (this.tracker) {
-      this.tracker.setFilter(filter);
-    }
-  }
-
-  /**
-   * Set the sort for tracker features.
-   * @param {Function} sort Sort function.
-   */
-  setSort(sort) {
-    if (this.tracker) {
-      this.tracker.setSort(sort);
-    }
-  }
-
-  getVehicle(filterFc) {
-    return this.tracker.getTrajectories().filter(filterFc);
-  }
-
-  /**
-   * Start to update the current time depending on the speed.
-   * @private
-   */
-  startUpdateTime() {
-    this.stopUpdateTime();
-    this.updateTime = setInterval(() => {
-      const newTime =
-        this.currTime.getTime() +
-        (new Date() - this.lastUpdateTime) * this.speed;
-      this.setCurrTime(newTime);
-    }, this.getRefreshTimeInMs());
-  }
-
-  /**
-   * Stop to update time.
-   * @private
-   */
-  stopUpdateTime() {
-    clearInterval(this.updateTime);
+    super.stop();
+    this.map.off('moveend', this.onMapMoveEnd);
+    this.map.off('mousemove', this.onMapMouseMove);
   }
 
   /**
@@ -279,82 +99,39 @@ class TrackerLayer extends Layer {
    */
   getVehiclesAtCoordinate(coordinate) {
     const resolution = getResolution(this.map);
-    const ext = buffer([...coordinate, ...coordinate], 10 * resolution);
-    const trajectories = this.tracker.getTrajectories();
-    const vehicles = [];
-    for (let i = 0; i < trajectories.length; i += 1) {
-      if (
-        trajectories[i].coordinate &&
-        containsCoordinate(ext, trajectories[i].coordinate)
-      ) {
-        vehicles.push(trajectories[i]);
-      }
-    }
-
-    return vehicles;
+    return super.getVehiclesAtCoordinate(coordinate, resolution);
   }
 
   /**
-   * Define the style of the vehicle.
-   * Draw a blue circle with the id of the props parameter.
-   *
-   * @param {Object} props Properties
    * @private
    */
-  style(props) {
-    const { id: text } = props;
-    if (this.styleCache[text]) {
-      return this.styleCache[text];
-    }
-    const canvas = document.createElement('canvas');
-    canvas.width = 200;
-    canvas.height = 15;
-    const ctx = canvas.getContext('2d');
-    ctx.arc(8, 8, 5, 0, 2 * Math.PI, false);
-    ctx.fillStyle = '#8ED6FF';
-    ctx.fill();
-    ctx.lineWidth = 3;
-    ctx.strokeStyle = 'black';
-    ctx.stroke();
-    ctx.font = 'bold 12px arial';
-    ctx.strokeStyle = 'white';
-    ctx.lineWidth = 3;
-    ctx.strokeText(text, 20, 10);
-    ctx.fillStyle = 'black';
-    ctx.fillText(text, 20, 10);
-    this.styleCache[text] = canvas;
-    return this.styleCache[text];
-  }
+  onMapMoveEnd() {
+    const z = this.map.getZoom();
 
-  /**
-   * Listens to click events on the layer.
-   * @param {function} callback Callback function, called with the clicked
-   *   features (https://openlayers.org/en/latest/apidoc/module-ol_Feature.html),
-   *   the layer instance and the click event.
-   */
-  onClick(callback) {
-    if (typeof callback === 'function') {
-      if (!this.clickCallbacks.includes(callback)) {
-        this.clickCallbacks.push(callback);
-      }
-    } else {
-      throw new Error('callback must be of type function.');
+    if (z !== this.currentZoom) {
+      this.currentZoom = z;
+      this.startUpdateTime(z);
     }
   }
 
   /**
-   * Unlistens to click events on the layer.
-   * @param {function} callback Callback function, called with the clicked
-   *   features (https://openlayers.org/en/latest/apidoc/module-ol_Feature.html),
-   *   the layer instance and the click event.
+   * @private
    */
-  unClick(callback) {
-    if (typeof callback === 'function') {
-      const idx = this.clickCallbacks.indexOf(callback);
-      if (idx >= -1) {
-        this.clickCallbacks.splice(idx, 1);
-      }
+  onMapMouseMove(evt) {
+    if (
+      this.map.isMoving() ||
+      this.map.isRotating() ||
+      this.map.isZooming() ||
+      !this.isHoverActive
+    ) {
+      this.map.getContainer().style.cursor = 'auto';
+      return;
     }
+    const [vehicle] = this.getVehiclesAtCoordinate(
+      fromLonLat([evt.lngLat.lng, evt.lngLat.lat]),
+    );
+    this.map.getContainer().style.cursor = vehicle ? 'pointer' : 'auto';
+    this.tracker.setHoverVehicleId(vehicle && vehicle.id);
   }
 }
